+++ CHANGES.md
@@ -1,5 +1,6 @@
 # Version 1.1.0
 
+- Add printer operator for `I2C::ScanResult`
 - Add `Pin::PulseScope` for using RAII to set and reset a pin
 - Add `Flash` class for accessing flash devices
 - Add additional methods to `Drive` class for easier access to attribute flags